% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestsDerivationFunctions.R
\name{FindSameItems}
\alias{FindSameItems}
\title{Find same items}
\usage{
FindSameItems(
  smr,
  threshold = 1,
  itemColumnName = smr$ItemColumnName,
  sampleSize = NULL
)
}
\arguments{
\item{smr}{An SMR object.}

\item{threshold}{Threshold above which two items (rows)
in the SMR are considered to be "the same".}

\item{itemColumnName}{A string specifying the column names in the result.
If NULL \code{smr$ItemColumnName} is used.}

\item{sampleSize}{A positive integer specifying a sample of items to be used.
If NULL all items \code{smr} are used.}
}
\value{
A data frame.
}
\description{
For each test in specified data frame finds the labels that are "the same"
according to the specified SMR object and similarity threshold,
and makes those "same" labels to have the same (recommendation) scores.
}
