% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestsDerivationFunctions.R
\name{MakeItemsetsRecommender}
\alias{MakeItemsetsRecommender}
\title{Make item-sets recommender}
\usage{
MakeItemsetsRecommender(itemsets)
}
\arguments{
\item{itemsets}{A character vector of "collapsed" frequent itemsets.}
}
\value{
An SMR object.
}
\description{
Creates item-sets recommender for a given list of item-sets.
}
