% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestsDerivationFunctions.R
\name{MakeTestByClassification}
\alias{MakeTestByClassification}
\title{Make a test by classification}
\usage{
MakeTestByClassification(smr, tagType, profile, nTestLabels, ...)
}
\arguments{
\item{smr}{An SMR object.}

\item{tagType}{A tag type. (A string).}

\item{profile}{A profile. Has a form accepted by \code{\link{SMRMonClassifyByProfile}}.}

\item{nTestLabels}{Number of test labels in the result.}

\item{...}{Additional arguments for \code{\link{SMRMonClassifyByProfile}}.}
}
\value{
A data frame.
}
\description{
Makes a classification test for given SMR object, tag type, and profile.
}
