% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestsDerivationFunctions.R
\name{MakeTestByRecommendation}
\alias{MakeTestByRecommendation}
\title{Make a test by recommendation}
\usage{
MakeTestByRecommendation(
  smr,
  profile,
  nrecs,
  tagType = smr$ItemColumnName,
  ...
)
}
\arguments{
\item{smr}{An SMR object.}

\item{profile}{A profile. Has a form accepted by \code{\link{SMRMonRecommendByProfile}}.}

\item{nrecs}{Number of recommendations.}

\item{tagType}{The tag type to assign to the test.}

\item{...}{Additional arguments for \code{\link{SMRMonRecommendByProfile}}.}
}
\value{
A data frame.
}
\description{
Makes a recommendation test for given SMR object and profile.
}
