% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestsDerivationFunctions.R
\name{RepresentByItemsets}
\alias{RepresentByItemsets}
\title{Represent by item-sets}
\usage{
RepresentByItemsets(smrItemsets, profile)
}
\arguments{
\item{smrItemsets}{An item-sets recommender.}

\item{profile}{A profile that has a form accepted by \code{\link{SMRMonRecommendByProfile}}.}
}
\value{
A data frame
}
\description{
Finds the representation a profile over the columns of
a given item-sets recommender.
}
