% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSTMon.R
\name{RSTMonEcho}
\alias{RSTMonEcho}
\title{Echo.}
\usage{
RSTMonEcho(rstObj, x)
}
\arguments{
\item{rstObj}{An RSTMon object.}

\item{x}{Object to be echoed.}
}
\value{
An RSTMon object
}
\description{
Echoes given argument.
}
\concept{Echo functions}
