% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSTMon.R
\name{RSTMonEchoFunctionValue}
\alias{RSTMonEchoFunctionValue}
\title{Echo function application to monad's value.}
\usage{
RSTMonEchoFunctionValue(rstObj, f)
}
\arguments{
\item{rstObj}{An RSTMon object.}

\item{f}{A function to be applied to \code{rstObj$Value}.}
}
\value{
A RSTMon object.
}
\description{
Applies a function to the "Value" element/member of the monad object
and prints the result.
}
\details{
Prints \code{f(rstObj$Value)}.
}
