% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSTMon.R
\name{RSTMonEchoValue}
\alias{RSTMonEchoValue}
\title{Echo monad's value.}
\usage{
RSTMonEchoValue(rstObj)
}
\arguments{
\item{rstObj}{An RSTMon object.}
}
\value{
A RSTMon object.
}
\description{
Prints the "Value" element/member of the monad object.
}
\details{
Prints \code{rstObj$Value}.
}
