% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSTMon.R
\name{RSTMonExportToCSVFile}
\alias{RSTMonExportToCSVFile}
\title{ExportToCSVFile}
\usage{
RSTMonExportToCSVFile(rstObj, fileName, ...)
}
\arguments{
\item{rstObj}{An RSTMon object.}

\item{fileName}{A string that is file name.}

\item{...}{Additional parameters for \code{\link{write.csv}}.}
}
\value{
An RSTMon object.
}
\description{
Exports the pipeline value to a given CSV file.
}
\seealso{
Other Export functions: 
\code{\link{RSTMonExportToFeatherFile}()}
}
\concept{Export functions}
