% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSTMon.R
\name{RSTMonExportToFeatherFile}
\alias{RSTMonExportToFeatherFile}
\title{ExportToFeatherFile.}
\usage{
RSTMonExportToFeatherFile(rstObj, fileName)
}
\arguments{
\item{rstObj}{An RSTMon object.}

\item{fileName}{A string that is file name.}
}
\value{
An RSTMon object.
}
\description{
Exports the pipeline value to a given feather file.
}
\seealso{
Other Export functions: 
\code{\link{RSTMonExportToCSVFile}()}
}
\concept{Export functions}
