% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSTMon.R
\name{RSTMonFailure}
\alias{RSTMonFailure}
\title{Failure producing function}
\usage{
RSTMonFailure(rstObj, message = NULL)
}
\arguments{
\item{rstObj}{An RSTMon object.}

\item{message}{A message to echoed. If NULL no message is echoed.}
}
\value{
A RSTMon object.
}
\description{
Puts the monad object into a monadic failure.
}
