% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSTMon.R
\docType{data}
\name{RSTMonFailureSymbol}
\alias{RSTMonFailureSymbol}
\title{Failure symbol for RSTMon.}
\format{
An object of class \code{logical} of length 1.
}
\usage{
RSTMonFailureSymbol
}
\description{
Failure symbol for the monad RSTMon.
}
\keyword{datasets}
