% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSTMon.R
\name{RSTMonFilterItemsets}
\alias{RSTMonFilterItemsets}
\title{FilterItemsets.}
\usage{
RSTMonFilterItemsets(
  rstObj,
  pattern = NULL,
  minSupport = 0,
  maxSupport = max(rstObj$Itemsets$Support),
  minCount = 0,
  maxCount = max(rstObj$Itemsets$Count),
  ...
)
}
\arguments{
\item{rstObj}{An RSTMon object.}

\item{pattern}{A pattern for grepl}

\item{minSupport}{Min support.}

\item{maxSupport}{Max support.}

\item{minCount}{Min count.}

\item{maxCount}{Max count.}

\item{...}{Additional arguments for \code{\link{grepl}}.}
}
\value{
An RSTMon object or \code{RSTMonFailureSymbol}.
}
\description{
Ingests the SMR matrix data in a given feather file.
}
\details{
The data frame \code{rstObj$Itemsets} is modified using \code{\link{dplyr::filter}}.
The result of the filtering is also assigned to \code{rstObj$Value}.
}
\seealso{
Other Computation functions: 
\code{\link{RSTMonFindFrequentSets}()},
\code{\link{RSTMonMakeClassificationTests}()},
\code{\link{RSTMonMakeRecommendationTests}()}
}
\concept{Computation functions}
