% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSTMon.R
\name{RSTMonFindFrequentSets}
\alias{RSTMonFindFrequentSets}
\title{FindFrequentSets}
\usage{
RSTMonFindFrequentSets(
  rstObj,
  profileTagTypes = NULL,
  supp = NULL,
  minlen = 2,
  maxlen = 5,
  target = "maximally frequent itemsets"
)
}
\arguments{
\item{rstObj}{An RSTMon object.}

\item{profileTagTypes}{Tag type use for making the profiles.
If NULL then \code{rstObj$ProfileTagTypes}.}

\item{supp}{Min support for the found frequent sets.
If NULL then \code{rstObj$FreqSetMinSupport}.}

\item{minlen}{Min length for the found frequent sets.
If NULL then \code{rstObj$FreqSetsMinLen}.}

\item{maxlen}{Max length for the found frequent sets.
If NULL then \code{rstObj$FreqSetsMaxLen}.}

\item{target}{Type of frequent set found.}

\item{fileName}{A string that is file name.}
}
\value{
An RSTMon object or \code{RSTMonFailureSymbol}.
}
\description{
Find frequent sets of tags for given profile
tag types and algorithm parameters.
}
\details{
The obtained data frame of frequent sets is assigned
to \code{rstObj$Value} and \code{rstObj$Itemsets}.
}
\seealso{
Other Computation functions: 
\code{\link{RSTMonMakeClassificationTests}()},
\code{\link{RSTMonMakeRecommendationTests}()}
}
\concept{Computation functions}
