% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSTMon.R
\name{RSTMonMakeClassificationTests}
\alias{RSTMonMakeClassificationTests}
\title{MakeClassificationTests}
\usage{
RSTMonMakeClassificationTests(rstObj, tagType = NULL, nTestLabels = 10)
}
\arguments{
\item{rstObj}{An RSTMon object.}

\item{tagType}{The tag type to assign to the test.}
}
\value{
An RSTMon object or \code{RSTMonFailureSymbol}.
}
\description{
Ingests the SMR matrix data in a given feather file.
}
\details{
The obtained data frame of tests is assigned to
\code{rstObj$Value}.
}
\seealso{
Other Computation functions: 
\code{\link{RSTMonFindFrequentSets}()},
\code{\link{RSTMonMakeRecommendationTests}()}
}
\concept{Computation functions}
