% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSTMon.R
\name{RSTMonOption}
\alias{RSTMonOption}
\title{Optional function application to monad's object.}
\usage{
RSTMonOption(rstObj, f)
}
\arguments{
\item{rstObj}{An RSTMon object.}

\item{f}{A function to be applied to the monad object.}
}
\value{
A RSTMon object.
}
\description{
If monadic failure is obtained from \code{f(rstObj)}
then returns the original \code{rstObj};
else returns the result of \code{f(rstObj)}.
}
\details{
In general \code{f} should return a monad object,
but that is not enforced.
}
