% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSTMon.R
\name{RSTMonSetFreqSetMinSupport}
\alias{RSTMonSetFreqSetMinSupport}
\title{Set FreqSetMinSupport.}
\usage{
RSTMonSetFreqSetMinSupport(rstObj, FreqSetMinSupport)
}
\arguments{
\item{rstObj}{An RSTMon object.}

\item{FreqSetMinSupport}{An object member to be set.}
}
\value{
An RSTMon object.
}
\description{
Sets FreqSetMinSupport into the monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{RSTMonSetFreqSetMinCount}()},
\code{\link{RSTMonSetFreqSetsMaxLen}()},
\code{\link{RSTMonSetFreqSetsMinLen}()},
\code{\link{RSTMonSetItemsets}()},
\code{\link{RSTMonSetLabelTagType}()},
\code{\link{RSTMonSetNTopLabels}()},
\code{\link{RSTMonSetNTopRecs}()},
\code{\link{RSTMonSetProfileTagTypes}()},
\code{\link{RSTMonSetSMR}()},
\code{\link{RSTMonSetValue}()},
\code{\link{RSTMonTakeFreqSetMinCount}()},
\code{\link{RSTMonTakeFreqSetMinSupport}()},
\code{\link{RSTMonTakeFreqSetsMaxLen}()},
\code{\link{RSTMonTakeFreqSetsMinLen}()},
\code{\link{RSTMonTakeItemsets}()},
\code{\link{RSTMonTakeLabelTagType}()},
\code{\link{RSTMonTakeNTopLabels}()},
\code{\link{RSTMonTakeNTopRecs}()},
\code{\link{RSTMonTakeProfileTagTypes}()},
\code{\link{RSTMonTakeSMR}()},
\code{\link{RSTMonTakeValue}()}
}
\concept{Set/Take functions}
