% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSTMon.R
\name{RSTMonSetItemsets}
\alias{RSTMonSetItemsets}
\title{Set Itemsets.}
\usage{
RSTMonSetItemsets(rstObj, Itemsets)
}
\arguments{
\item{rstObj}{An RSTMon object.}

\item{Itemsets}{An object member to be set.}
}
\value{
An RSTMon object.
}
\description{
Sets Itemsets into the monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{RSTMonSetFreqSetMinCount}()},
\code{\link{RSTMonSetFreqSetMinSupport}()},
\code{\link{RSTMonSetFreqSetsMaxLen}()},
\code{\link{RSTMonSetFreqSetsMinLen}()},
\code{\link{RSTMonSetLabelTagType}()},
\code{\link{RSTMonSetNTopLabels}()},
\code{\link{RSTMonSetNTopRecs}()},
\code{\link{RSTMonSetProfileTagTypes}()},
\code{\link{RSTMonSetSMR}()},
\code{\link{RSTMonSetValue}()},
\code{\link{RSTMonTakeFreqSetMinCount}()},
\code{\link{RSTMonTakeFreqSetMinSupport}()},
\code{\link{RSTMonTakeFreqSetsMaxLen}()},
\code{\link{RSTMonTakeFreqSetsMinLen}()},
\code{\link{RSTMonTakeItemsets}()},
\code{\link{RSTMonTakeLabelTagType}()},
\code{\link{RSTMonTakeNTopLabels}()},
\code{\link{RSTMonTakeNTopRecs}()},
\code{\link{RSTMonTakeProfileTagTypes}()},
\code{\link{RSTMonTakeSMR}()},
\code{\link{RSTMonTakeValue}()}
}
\concept{Set/Take functions}
