% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSTMon.R
\name{RSTMonTakeNTopRecs}
\alias{RSTMonTakeNTopRecs}
\title{Take NTopRecs.}
\usage{
RSTMonTakeNTopRecs(rstObj, functionName = "RSTMonTakeNTopRecs")
}
\arguments{
\item{rstObj}{An RSTMon object.}

\item{functionName}{A string that is a name of this function or a delegating function.}
}
\value{
A list of functions or \code{RSTMonFailureSymbol}.
}
\description{
Takes NTopRecs from the monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{RSTMonSetFreqSetMinCount}()},
\code{\link{RSTMonSetFreqSetMinSupport}()},
\code{\link{RSTMonSetFreqSetsMaxLen}()},
\code{\link{RSTMonSetFreqSetsMinLen}()},
\code{\link{RSTMonSetItemsets}()},
\code{\link{RSTMonSetLabelTagType}()},
\code{\link{RSTMonSetNTopLabels}()},
\code{\link{RSTMonSetNTopRecs}()},
\code{\link{RSTMonSetProfileTagTypes}()},
\code{\link{RSTMonSetSMR}()},
\code{\link{RSTMonSetValue}()},
\code{\link{RSTMonTakeFreqSetMinCount}()},
\code{\link{RSTMonTakeFreqSetMinSupport}()},
\code{\link{RSTMonTakeFreqSetsMaxLen}()},
\code{\link{RSTMonTakeFreqSetsMinLen}()},
\code{\link{RSTMonTakeItemsets}()},
\code{\link{RSTMonTakeLabelTagType}()},
\code{\link{RSTMonTakeNTopLabels}()},
\code{\link{RSTMonTakeProfileTagTypes}()},
\code{\link{RSTMonTakeSMR}()},
\code{\link{RSTMonTakeValue}()}
}
\concept{Set/Take functions}
