% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSTMon.R
\name{RSTMonTakeValue}
\alias{RSTMonTakeValue}
\title{Take the value in a RSTMon object.}
\usage{
RSTMonTakeValue(rstObj)
}
\arguments{
\item{rstObj}{An RSTMon object.}
}
\value{
Just \code{rstObj$Value}.
}
\description{
Takes the value from RSTMon monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{RSTMonSetFreqSetMinCount}()},
\code{\link{RSTMonSetFreqSetMinSupport}()},
\code{\link{RSTMonSetFreqSetsMaxLen}()},
\code{\link{RSTMonSetFreqSetsMinLen}()},
\code{\link{RSTMonSetItemsets}()},
\code{\link{RSTMonSetLabelTagType}()},
\code{\link{RSTMonSetNTopLabels}()},
\code{\link{RSTMonSetNTopRecs}()},
\code{\link{RSTMonSetProfileTagTypes}()},
\code{\link{RSTMonSetSMR}()},
\code{\link{RSTMonSetValue}()},
\code{\link{RSTMonTakeFreqSetMinCount}()},
\code{\link{RSTMonTakeFreqSetMinSupport}()},
\code{\link{RSTMonTakeFreqSetsMaxLen}()},
\code{\link{RSTMonTakeFreqSetsMinLen}()},
\code{\link{RSTMonTakeItemsets}()},
\code{\link{RSTMonTakeLabelTagType}()},
\code{\link{RSTMonTakeNTopLabels}()},
\code{\link{RSTMonTakeNTopRecs}()},
\code{\link{RSTMonTakeProfileTagTypes}()},
\code{\link{RSTMonTakeSMR}()}
}
\concept{Set/Take functions}
