% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSTMon.R
\name{RSTMonUnit}
\alias{RSTMonUnit}
\title{Make a RSTMon Unit}
\usage{
RSTMonUnit()
}
\value{
An S3 class "RSTMon". In other words, a list with the attribute "class" set to "RSTMon".
}
\description{
Creates a monad object.
}
