% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComparisonFunctions.R
\name{SummarizeRanks}
\alias{SummarizeRanks}
\title{Summarize ranks}
\usage{
SummarizeRanks(data, focusTags = NULL, summariseByTagTypeQ = FALSE)
}
\arguments{
\item{data}{A data frame.}

\item{focusTags}{A character vector with tags over which the summaries are computed.}

\item{summariseByTagTypeQ}{Should the tag-summaries be further summarized by tag type or not?}

\item{A}{data frame.}
}
\description{
Summarizes the ranks of given data frame.
}
