% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComparisonFunctions.R
\name{TestResultsComparisonMetrics}
\alias{TestResultsComparisonMetrics}
\title{Test results comparison metrics}
\usage{
TestResultsComparisonMetrics(
  result1,
  result2,
  nTop = 10,
  by = c("QueryType", "Query", "TagType"),
  tagColumnName = "Tag",
  corMethod = "pearson"
)
}
\arguments{
\item{result1}{First recommendations set.}

\item{result2}{Second recommendations set.}

\item{nTop}{The top number of recommendations to take.}

\item{by}{A character vector with column names to group by.
If NULL no grouping is made.}

\item{tagColumnName}{The column that has the recommendation tags.}

\item{corMethod}{Method for \code{\link{cor}}.
(One of \code{c("pearson", "kendall", "spearman")}.)}
}
\value{
A data frame.
}
\description{
Compares two recommendations results data frames.
Intersection size and correlation of scores are computed.
}
